requirements.txt
appdirs==1.4.3
attrs==18.2.0
boto==2.49.0
boto3==1.9.106
botocore==1.12.106
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1.7
django-storages==1.7.1
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
Markdown==3.0.1
psycopg2==2.7.7
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
toml==0.10.0
urllib3==1.24.1
